# (2.12) Optimised meta snapshot encoding when `GOEXPERIMENT=jsonv2` enabled