With the old JSON library:
```
go test -v ./server -run=XXX -bench=BenchmarkJetStreamMetaSnapshot -benchtime=1x
goos: darwin
goarch: arm64
pkg: github.com/nats-io/nats-server/v2/server
cpu: Apple M2 Ultra
BenchmarkJetStreamMetaSnapshot
BenchmarkJetStreamMetaSnapshot/Default
BenchmarkJetStreamMetaSnapshot/Default-24         	       1	 466903916 ns/op
BenchmarkJetStreamMetaSnapshot/AllUnsupported
BenchmarkJetStreamMetaSnapshot/AllUnsupported-24  	       1	 267354584 ns/op
PASS
ok  	github.com/nats-io/nats-server/v2/server	9.826s
```

With the new JSON library:
```
GOEXPERIMENT=jsonv2 go test -v ./server -run=XXX -bench=BenchmarkJetStreamMetaSnapshot -benchtime=1x
goos: darwin
goarch: arm64
pkg: github.com/nats-io/nats-server/v2/server
cpu: Apple M2 Ultra
BenchmarkJetStreamMetaSnapshot
BenchmarkJetStreamMetaSnapshot/Default
BenchmarkJetStreamMetaSnapshot/Default-24         	       1	 196555083 ns/op
BenchmarkJetStreamMetaSnapshot/AllUnsupported
BenchmarkJetStreamMetaSnapshot/AllUnsupported-24  	       1	 205764542 ns/op
PASS
ok  	github.com/nats-io/nats-server/v2/server	9.154s
```

Signed-off-by: Neil Twigg <neil@nats.io>